rename typing.py to custom_types.py for clarity
