+++ libp2p/identity/identify/protocol.py
@@ -4,6 +4,10 @@ from multiaddr import (
     Multiaddr,
 )
 
+from libp2p.custom_types import (
+    StreamHandlerFn,
+    TProtocol,
+)
 from libp2p.host.host_interface import (
     IHost,
 )
@@ -12,10 +16,6 @@ from libp2p.network.stream.exceptions import (
 )
 from libp2p.network.stream.net_stream_interface import (
     INetStream,
-)
-from libp2p.typing import (
-    StreamHandlerFn,
-    TProtocol,
 )
 
 from .pb.identify_pb2 import (
