+++ libp2p/network/stream/net_stream.py
@@ -2,6 +2,9 @@ from typing import (
     Optional,
 )
 
+from libp2p.custom_types import (
+    TProtocol,
+)
 from libp2p.stream_muxer.abc import (
     IMuxedStream,
 )
@@ -9,9 +12,6 @@ from libp2p.stream_muxer.exceptions import (
     MuxedStreamClosed,
     MuxedStreamEOF,
     MuxedStreamReset,
-)
-from libp2p.typing import (
-    TProtocol,
 )
 
 from .exceptions import (
