+++ libp2p/protocol_muxer/multiselect_client.py
@@ -2,7 +2,7 @@ from collections.abc import (
     Sequence,
 )
 
-from libp2p.typing import (
+from libp2p.custom_types import (
     TProtocol,
 )
