+++ libp2p/protocol_muxer/multiselect_muxer_interface.py
@@ -3,7 +3,7 @@ from abc import (
     abstractmethod,
 )
 
-from libp2p.typing import (
+from libp2p.custom_types import (
     StreamHandlerFn,
     TProtocol,
 )
