+++ libp2p/stream_muxer/muxer_multistream.py
@@ -2,6 +2,9 @@ from collections import (
     OrderedDict,
 )
 
+from libp2p.custom_types import (
+    TProtocol,
+)
 from libp2p.network.connection.raw_connection_interface import (
     IRawConnection,
 )
@@ -23,9 +26,6 @@ from libp2p.security.secure_conn_interface import (
 from libp2p.transport.typing import (
     TMuxerClass,
     TMuxerOptions,
-)
-from libp2p.typing import (
-    TProtocol,
 )
 
 from .abc import (
