+++ libp2p/tools/factories.py
@@ -28,6 +28,9 @@ from libp2p.crypto.keys import (
     PrivateKey,
 )
 from libp2p.crypto.secp256k1 import create_new_key_pair as create_secp256k1_key_pair
+from libp2p.custom_types import (
+    TProtocol,
+)
 from libp2p.host.basic_host import (
     BasicHost,
 )
@@ -120,9 +123,6 @@ from libp2p.transport.typing import (
 )
 from libp2p.transport.upgrader import (
     TransportUpgrader,
-)
-from libp2p.typing import (
-    TProtocol,
 )
 
 from .constants import (
