+++ newsfragments/500.internal.rst
@@ -1,0 +1,1 @@
+Rename ``typing.py`` to ``custom_types.py`` for clarity.
